cleanups, splitting tests across files
