+++ build.sbt
@@ -93,12 +93,13 @@ lazy val interpreter = (project in file("interpreter"))
     parallelExecution in Benchmark := false
   )
 
-val akkaActor   = "com.typesafe.akka" %% "akka-actor"           % "2.5.23"
-val akkaStream  = "com.typesafe.akka" %% "akka-stream"          % "2.5.23"
-val akkaHttp    = "com.typesafe.akka" %% "akka-http"            % "10.1.8"
-val akkaSpray   = "com.typesafe.akka" %% "akka-http-spray-json" % "10.1.8"
-val akkaTyped   = "com.typesafe.akka" %% "akka-actor-typed"     % "2.5.23"
-val akkaTestkit = "com.typesafe.akka" %% "akka-testkit"         % "2.5.23"
+val akkaActor        = "com.typesafe.akka" %% "akka-actor"               % "2.5.23"
+val akkaStream       = "com.typesafe.akka" %% "akka-stream"              % "2.5.23"
+val akkaHttp         = "com.typesafe.akka" %% "akka-http"                % "10.1.8"
+val akkaSpray        = "com.typesafe.akka" %% "akka-http-spray-json"     % "10.1.8"
+val akkaTyped        = "com.typesafe.akka" %% "akka-actor-typed"         % "2.5.23"
+val akkaTestkit      = "com.typesafe.akka" %% "akka-testkit"             % "2.5.23"
+val akkaSLF4J        = "com.typesafe.akka" %% "akka-slf4j"               % "2.5.23"
 val akkaTestkitTyped = "com.typesafe.akka" %% "akka-actor-testkit-typed" % "2.5.23" % Test
 
 val akka = Seq(akkaActor, akkaStream, akkaHttp, akkaSpray, akkaTyped)
@@ -113,10 +114,12 @@ lazy val fileManager = (project in file("file-manager"))
   )
   .settings(
     libraryDependencies ++= akka,
-    libraryDependencies += "org.scalatest"  %% "scalatest"  % "3.2.0-SNAP10" % Test,
-    libraryDependencies += "org.scalacheck" %% "scalacheck" % "1.14.0" % Test,
+    libraryDependencies += akkaSLF4J,
+    libraryDependencies += "ch.qos.logback" % "logback-classic" % "1.2.3",
+    libraryDependencies += "org.scalatest"  %% "scalatest"      % "3.2.0-SNAP10" % Test,
+    libraryDependencies += "org.scalacheck" %% "scalacheck"     % "1.14.0" % Test,
     libraryDependencies += akkaTestkitTyped,
-    libraryDependencies += "commons-io" % "commons-io" % "2.6",
+    libraryDependencies += "commons-io" % "commons-io"        % "2.6",
     libraryDependencies += "io.methvin" % "directory-watcher" % "0.9.6"
   )
