+++ file-manager/src/test/scala/org/enso/filemanager/FileManagerBehaviorTests.scala
@@ -41,69 +41,14 @@ import scala.concurrent.Await
 import scala.concurrent.Future
 import scala.reflect.ClassTag
 import scala.concurrent.duration.FiniteDuration
-import scala.concurrent.duration._
 import scala.util.Failure
 import scala.util.Success
 import scala.util.Try
 
-trait FileSystemHelpers {
-  var tempDir: Path = _
-
-  val contents: Array[Byte] =
-    "葦垣の中の和草にこやかに我れと笑まして人に知らゆな\nzażółć gęślą jaźń".getBytes
-
-  def createSubFile(): Path = {
-    val path = Files.createTempFile(tempDir, "foo", "")
-    Files.write(path, contents)
-  }
-
-  def createSubDir(): Path = {
-    Files.createTempDirectory(tempDir, "foo")
-  }
-
-  def homeDirectory(): Path = Paths.get(System.getProperty("user.home"))
-
-  def setupTemp(): Unit = {
-    tempDir = Files.createTempDirectory("file-manager-test")
-  }
-
-  def cleanTemp(): Unit = {
-    FileUtils.deleteDirectory(tempDir.toFile)
-    tempDir = null
-  }
-
-  def withTemporaryDirectory[ret](f: Path => ret): ret = {
-    setupTemp()
-    try f(tempDir)
-    finally cleanTemp()
-  }
-
-  case class Subtree(
-    root: Path,
-    childrenFiles: Seq[Path],
-    childrenDirs: Seq[Path]) {
-
-    val elements: Seq[Path] =
-      (Seq(root) ++ childrenDirs ++ childrenFiles).map(root.resolve(_))
-
-    def rebase(otherRoot: Path): Subtree =
-      Subtree(otherRoot, childrenFiles, childrenDirs)
-  }
-
-  def createSubtree(): Subtree = {
-    val root       = createSubDir()
-    val rootFile1  = Paths.get("file1")
-    val rootSubDir = Paths.get("dir")
-    val rootFile2  = Paths.get("dir/file2")
-
-    Files.write(root.resolve(rootFile1), contents)
-    Files.createDirectory(root.resolve(rootSubDir))
-    Files.write(root.resolve(rootFile2), contents)
-    Subtree(root, Seq(rootFile1, rootFile2), Seq(rootSubDir))
-  }
-}
-
-class FileManagerTests extends FunSuite with Matchers with FileSystemHelpers {
+class FileManagerBehaviorTests
+    extends FunSuite
+    with Matchers
+    with FileSystemHelpers {
   import API._
 
   var testKit: BehaviorTestKit[InputMessage] = _
@@ -135,25 +80,6 @@ class FileManagerTests extends FunSuite with Matchers with FileSystemHelpers {
       case Success(msg) =>
         fail(s"Unexpected non-error message: $msg")
     }
-  }
-
-  def expectSubtree(subtree: Subtree): Unit = {
-    assert(Files.exists(subtree.root))
-    subtree.elements.foreach(
-      elem => expectExist(subtree.root.resolve(elem))
-    )
-
-    val listStream = Files.list(subtree.root)
-    try listStream.count() should be(2)
-    finally listStream.close()
-  }
-
-  def expectExist(path: Path): Unit = {
-    assert(Files.exists(path), s"$path is expected to exist")
-  }
-
-  def expectNotExist(path: Path): Unit = {
-    assert(!Files.exists(path), s"$path is expected to not exist")
   }
 
   def runRequest(contents: RequestPayload[SuccessResponse]): Unit =
@@ -402,133 +328,4 @@ class FileManagerTests extends FunSuite with Matchers with FileSystemHelpers {
   }
 }
 
-// needs to be separate because watcher message are asynchronous
-class FileManagerWatcherTests
-    extends FunSuite
-    with BeforeAndAfterAll
-    with Matchers
-    with FileSystemHelpers {
-  import API._
 
-  var testKit: ActorTestKit         = ActorTestKit()
-  implicit val timeout: Timeout     = 3.seconds
-  implicit val scheduler: Scheduler = testKit.scheduler
-
-  var fileManager: ActorRef[InputMessage]   = _
-  var testProbe: TestProbe[FileSystemEvent] = _
-  var watcherID: UUID                       = _
-
-  override def withFixture(test: NoArgTest): Outcome = {
-    withTemporaryDirectory(_ => {
-      fileManager = testKit.spawn(FileManager.fileManager(tempDir))
-      testProbe   = testKit.createTestProbe[FileSystemEvent]("file-observer")
-      watcherID   = observe(tempDir)
-      super.withFixture(test)
-    })
-  }
-
-  override def afterAll() {
-    testKit.shutdownTestKit()
-  }
-
-  def matchesEvent(
-    path: Path,
-    eventType: DirectoryChangeEvent.EventType
-  ): FileSystemEvent => Boolean = { message: FileSystemEvent =>
-    message.event.path() == path && message.event.eventType() == eventType
-  }
-
-  def expectEventPresentIn(
-    path: Path,
-    eventType: DirectoryChangeEvent.EventType,
-    events: Seq[FileSystemEvent]
-  ): Unit = {
-    assert(
-      events.exists(matchesEvent(path, eventType)),
-      s"not received message about $path"
-    )
-  }
-
-  def expectNextEvent(
-    path: Path,
-    eventType: DirectoryChangeEvent.EventType
-  ): Unit = {
-    val message = testProbe.receiveMessage()
-    assert(
-      matchesEvent(path, eventType)(message),
-      s"expected of type $eventType for $path, got $message"
-    )
-  }
-
-  def ask[response <: SuccessResponse: ClassTag](
-    requestPayload: RequestPayload[response]
-  ): Future[Try[response]] = {
-    val futureResponse = fileManager.ask(
-      (replyTo: ActorRef[Try[response]]) => {
-        val request =
-          Request(replyTo, requestPayload)
-        request: InputMessage
-      }
-    )
-    futureResponse
-  }
-
-  def observe(path: Path): UUID = {
-    val futureResponse = ask(CreateWatcherRequest(path, testProbe.ref))
-    Await.result(futureResponse, 50.millis).get.id
-  }
-
-  test("Watcher: observe subtree creation and deletion") {
-    val subtree = createSubtree()
-    val events  = testProbe.receiveMessages(subtree.elements.size)
-    subtree.elements.foreach(
-      expectEventPresentIn(_, DirectoryChangeEvent.EventType.CREATE, events)
-    )
-
-    FileUtils.deleteDirectory(subtree.root.toFile)
-
-    val deletionEvents = testProbe.receiveMessages(subtree.elements.size)
-    subtree.elements.foreach(
-      expectEventPresentIn(
-        _,
-        DirectoryChangeEvent.EventType.DELETE,
-        deletionEvents
-      )
-    )
-
-    testProbe.expectNoMessage(50.millis)
-  }
-
-  test("Watcher: observe file modification") {
-    // should generate two events
-    val dir10 = tempDir.resolve("dir10")
-    val dir20 = dir10.resolve("dir20")
-    // create two directories at once - we should get two notifications
-    Files.createDirectories(dir20)
-    expectNextEvent(dir10, DirectoryChangeEvent.EventType.CREATE)
-    expectNextEvent(dir20, DirectoryChangeEvent.EventType.CREATE)
-
-    val someFile = dir20.resolve("file.dat")
-    Files.createFile(someFile)
-    expectNextEvent(someFile, DirectoryChangeEvent.EventType.CREATE)
-    Files.write(someFile, "blahblah".getBytes)
-    expectNextEvent(someFile, DirectoryChangeEvent.EventType.MODIFY)
-
-    // deleting dir removes the file first
-    FileUtils.deleteDirectory(dir20.toFile)
-    expectNextEvent(someFile, DirectoryChangeEvent.EventType.DELETE)
-    expectNextEvent(dir20, DirectoryChangeEvent.EventType.DELETE)
-    testProbe.expectNoMessage(50.millis)
-  }
-
-  test("Watcher: disabling watch") {
-    val subtree = createSubtree()
-    testProbe.receiveMessages(subtree.elements.size)
-    val stopResponse =
-      Await.result(ask(WatcherRemoveRequest(watcherID)), 1.second)
-    stopResponse should be(Success(WatcherRemoveResponse()))
-
-    FileUtils.deleteDirectory(subtree.root.toFile)
-    testProbe.expectNoMessage(50.millis)
-  }
-}
